feat: clean come console.log, split code for highter readability
